backend/requirements.txt
asgiref==3.8.1
cffi==1.17.1
cryptography==44.0.2
Django==5.1.7
django-cors-headers==4.7.0
djangorestframework==3.15.2
mysqlclient==2.2.7
pycparser==2.22
sqlparse==0.5.3
